Fix dtype mismatch in TabrModel forward pass

+++ pytabkit/models/nn_models/tabr.py
@@ -388,7 +388,7 @@ class TabrModel(nn.Module):
         probs = F.softmax(similarities, dim=-1)
         probs = self.dropout(probs)
 
-        context_y_emb = self.label_encoder(candidate_y[context_idx][..., None])
+        context_y_emb = self.label_encoder(candidate_y[context_idx][..., None].float())
         values = context_y_emb + self.T(k[:, None] - context_k)
         context_x = (probs[:, None] @ values).squeeze(1)
         x = x + context_x
